discord_bot/requirements.txt
discord.py==2.2.3
python-dotenv==0.20.0
requests==2.31.0